Pass openssl 1.0.2 through to python3.5 override (#4)
It seems specifically Python 3.5.2 and earlier are not compatible with openssl > 1.0.2: patch. Because of how nix is all-wired up here, it uses whatever the latest version of openssl is which ends up being 1.1.0. Perfectly fine for later versions of python but not the one we want

Thankfully we can pass openssl 1.0.2 through to the Python compilation steps which seems to work. This requires setting an insecure package environment variable temporarily but afterwards works just fine.